Open install data flow directly
