+++ App.axaml.cs
@@ -20,10 +20,6 @@ namespace ATC4_HQ
                 desktop.MainWindow = new MainWindow
                 {
                     // 设置 MainWindow 的 DataContext
-                    // 假设 MainWindowViewModel 会包含 InstallGameViewModel
-                    // 或者 MainWindow 的内容就是 InstallGame UserControl
-                    // 这里我们假设 MainWindow 的 DataContext 是 MainWindowViewModel
-                    // 并且 MainWindowViewModel 会实例化 InstallGameViewModel
                     DataContext = new MainWindowViewModel() 
                 };
             }
