+++ Views/InstallGameData.axaml.cs
@@ -4,7 +4,6 @@ using Avalonia.Interactivity;
 using System.Threading.Tasks;
 using ATC4_HQ.ViewModels;
 using System; // 添加此 using 引用，用于 EventArgs 和 Console
-using Microsoft.Extensions.Logging;
 
 namespace ATC4_HQ.Views
 {
@@ -23,6 +22,8 @@ namespace ATC4_HQ.Views
             {
                 viewModel.RequestOpenFilePicker += OnRequestOpenFilePicker;
                 viewModel.RequestSaveFileDialog += OnRequestSaveFileDialog; // 添加保存文件对话框事件处理
+                viewModel.InstallGameDataCompleted += OnInstallGameDataCompleted;
+                viewModel.ClearSubPageRequested += OnClearSubPageRequested;
             }
         }
 
@@ -32,6 +33,8 @@ namespace ATC4_HQ.Views
             {
                 viewModel.RequestOpenFilePicker -= OnRequestOpenFilePicker;
                 viewModel.RequestSaveFileDialog -= OnRequestSaveFileDialog; // 取消订阅保存文件对话框事件
+                viewModel.InstallGameDataCompleted -= OnInstallGameDataCompleted;
+                viewModel.ClearSubPageRequested -= OnClearSubPageRequested;
             }
         }
 
@@ -50,6 +53,37 @@ namespace ATC4_HQ.Views
             {
                 string? selectedPath = await OpenSaveFileDialog();
                 e.SetResult(selectedPath);
+            }
+        }
+
+        private async void OnInstallGameDataCompleted(object? sender, InstallGameDataCompletedEventArgs e)
+        {
+            if (e.Success && e.GameData != null)
+            {
+                LoggerHelper.LogInformation($"获取到的游戏数据 - 名称: {e.GameData.Name}, 路径: {e.GameData.Path}");
+
+                if (TopLevel.GetTopLevel(this) is Window mainWindow && mainWindow.DataContext is MainWindowViewModel mainWindowViewModel)
+                {
+                    await mainWindowViewModel.HandleInstallGameAndUnzipAsync(e.GameData);
+                    LoggerHelper.LogInformation("游戏数据已成功传递给 MainWindowViewModel 进行处理。");
+                }
+                else
+                {
+                    LoggerHelper.LogError("错误：无法获取 MainWindowViewModel 来处理游戏数据。");
+                }
+            }
+            else
+            {
+                LoggerHelper.LogInformation("用户取消了安装或安装失败。");
+            }
+        }
+
+        private void OnClearSubPageRequested(object? sender, EventArgs e)
+        {
+            if (TopLevel.GetTopLevel(this) is Window mainWindow && mainWindow.DataContext is MainWindowViewModel mainWindowViewModel)
+            {
+                mainWindowViewModel.ClearSubPage();
+                LoggerHelper.LogInformation("已清除右边区域的内容。");
             }
         }
